Fix starship builds
builds were failing on amd64 hosts due to xargs realpath missing it's operand with the following error:

```
realpath: missing operand
```

The fix for this is relatively simple, add -I "{}" back in to the xargs options, and then add "{}" to the realpath.

Further, the aarch64 builds were failing because it was downloading file `starship-aarch64-apple-darwin.tar.gz`. Macos was listed in the ignored file names list, but that's not in the name and apple and darwin were still allowed, so I've added them to the ignore list.